pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["erkam <46300684+erkamguresen@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.2"
Flask-HTTPAuth = "^4.7.0"
PyJWT = "^2.4.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.3"
marshmallow-sqlalchemy = "^0.28.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
